TMV#84 Added pagination of the tasks list

+++ TimeTracker.Api.Shared/Dto/RequestsAndResponses/Dashboard/Tasks/GetListRequest.cs
@@ -12,7 +12,7 @@ namespace TimeTracker.Api.Shared.Dto.RequestsAndResponses.Dashboard.Tasks
 
         [Required]
         [IsPositive]
-        public long Page { get; set; } = 1;
+        public int Page { get; set; } = 1;
         
         public GetListFilterRequest? Filter { get; set; }
     }
